Separate to type.py and types.py
